+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [1.10.3] - Unreleased
 ### Fixed
 - Improved `PaginateResult` type to support Index Signature `unknown`.
+- Improved `Paginator` type generic.
 - Fixed `expired` middleware to remove the charset from the `Content-Type` header [#233].
 
 ## [1.10.2] - 2022-08-11